Update ReflowOven.ino
To handle millis() function overflow the difference is casted to uint32_t.

+++ examples/ReflowOven/ReflowOven.ino
@@ -39,7 +39,7 @@ void loop()
 {
     currentMillis = millis();
 
-    if ((currentMillis - previousMillis) >= interval)
+    if ((uint32_t)(currentMillis - previousMillis) >= interval)
     {
         previousMillis = currentMillis;
         ovenTemperature = thermocouple.readCelsius();
